Modified input_output folder structure
Our base class for DataIO is placed within the input_output folder at the same level as the handlers folder.

The handler folder contains all subclasses that inherit from the base class (i.e., "csvIO")

I added empty __init__.py files so that the directory is recognized as a python package and it can be imported from other places. Right now the __init__ files are empty, but in the future we could use it to set up logging, initialize package-level variables, etc.