Created configuration directory
Created a folder to house the configuration python file and yaml files of each model. The configuration python file ("config_manager.py") deals with figuring out what data source and columns to read based on a specified model. This output is passed to DataIO class so it can read the data from file