Add DataIO Abstract Base Class
Created the abstract base class for the DataIO module.

Instead of taking path and list of columns as input, it only takes a dictionary with the model's configuration. The dictionary contains all of the sources, columns of each source, and how to properly join each source to create a final df.

The read_df_from_path method should worry about reading and joining data.