Updated use of primary key throughout the code base
Previously we were setting gameid as the df index prior to merging. However, this was excluding the gameid column from the merged df. Because we are only dealing with two df for now, I updated the code base so that primary key, or join key, is a class variable part of dataset generator and not part of the source object. I also updated the yaml to reflect this