Updated & Finished unit test
The unittest package I'm using requires test files to start with "test_", so i changed the unittest file name.

I finished the unit test. What a fkn journey. Hated doing it but it did save us from some errors so im glad we decided to do it.

If you want you can read the 5 test cases i added. they are not that comprehensive but at least test the general logic of dataset generator.